refactor(cron): split automation render helpers
