+++ ui/src/pages/cron/view.ts
@@ -23,13 +23,12 @@ import type {
   CronJobsSortBy,
   CronSortDir,
 } from "../../api/types.ts";
-import { renderChannelPicker, type ChannelPickerOption } from "../../components/channel-picker.ts";
+import type { ChannelPickerOption } from "../../components/channel-picker.ts";
 import { renderCronJobsPagination } from "../../components/cron-jobs-pagination.ts";
 import { icon, icons } from "../../components/icons.ts";
 import { highlightCodeHtml } from "../../components/markdown-code-blocks.ts";
 import { renderModelPicker } from "../../components/model-picker.ts";
 import { providerIdFromModelRef } from "../../components/provider-icon.ts";
-import { renderPicker } from "../../components/select-picker.ts";
 import "../../components/tooltip.ts";
 import "../../components/web-awesome.ts";
 import "../../components/web-awesome-popover.ts";
@@ -38,7 +37,6 @@ import {
   renderSettingsRow,
   renderSettingsSection,
   renderSettingsToggle,
-  renderSettingsToggleRow,
 } from "../../components/settings-ui.ts";
 import { t } from "../../i18n/index.ts";
 import {
@@ -61,6 +59,15 @@ import { resolveScrollBehavior } from "../../lib/scroll-behavior.ts";
 import { renderSegmented } from "./segmented-control.ts";
 import { CRON_SUGGESTIONS, suggestionFormPatch } from "./suggestions.ts";
 import { renderJobsFilterPopover } from "./view-filters.ts";
+import {
+  errorIdForField,
+  renderCronInput,
+  renderCronInputField,
+  renderCronSelect,
+  renderCronSelectField,
+  renderFieldRow,
+  renderToggleRow,
+} from "./view-form-fields.ts";
 import { renderRunsSection, runStatusLabel } from "./view-runs.ts";
 
 type CronPanelMode = "overview" | "create" | "job";
@@ -73,7 +80,7 @@ export type CronListTab = "tasks" | "activity";
 export type CronDetailTab = "settings" | "history";
 type CronOverviewTab = CronJobsEnabledFilter | "activity";
 
-type CronProps = {
+export type CronProps = {
   basePath: string;
   agentId: string;
   loading: boolean;
@@ -211,10 +218,6 @@ const CRON_FIELD_LABEL_KEYS: Record<CronFieldKey, string> = {
   failureAlertCooldownSeconds: "cron.form.failureAlertCooldown",
 };
 
-function errorIdForField(key: CronFieldKey) {
-  return `cron-error-${key}`;
-}
-
 function inputIdForField(key: string) {
   return `cron-${key.replace(/[A-Z]/g, (letter) => `-${letter.toLowerCase()}`)}`;
 }
@@ -262,173 +265,6 @@ function focusFormField(id: string) {
     el.scrollIntoView({ block: "center", behavior: resolveScrollBehavior() });
   }
   el.focus();
-}
-
-function renderFieldError(message?: string, id?: string) {
-  if (!message) {
-    return nothing;
-  }
-  return html`<div id=${ifDefined(id)} class="cron-help cron-error">${t(message)}</div>`;
-}
-
-// ── Row primitives (settings design language) ──
-
-function renderRequiredTitle(label: string) {
-  return html`
-    ${label}
-    <span class="cron-required-marker" aria-hidden="true">*</span>
-    <span class="cron-required-sr">${t("cron.form.requiredSr")}</span>
-  `;
-}
-
-// Settings row whose control keeps its own validation message underneath. Mirrors
-// renderSettingsRow markup; local only so the title can be a real <label for> that gives the
-// wrapped control its accessible name (including the visually-hidden required marker).
-function renderFieldRow(params: {
-  label: string;
-  // Blank when the control is not labelable (e.g. a code block); the label then
-  // has no `for` target and the control carries its own aria-label.
-  controlId: string;
-  control: unknown;
-  required?: boolean;
-  help?: string;
-  error?: string;
-  errorId?: string;
-  stacked?: boolean;
-  wide?: boolean;
-}) {
-  const controlClass = params.wide ? "cron-control cron-control--wide" : "cron-control";
-  const control = params.error
-    ? html`<div class=${controlClass}>
-        ${params.control}${renderFieldError(params.error, params.errorId)}
-      </div>`
-    : html`<div class=${controlClass}>${params.control}</div>`;
-  return html`
-    <div class=${params.stacked ? "settings-row settings-row--stacked" : "settings-row"}>
-      <label class="settings-row__text" for=${ifDefined(params.controlId || undefined)}>
-        <span class="settings-row__title">
-          ${params.required ? renderRequiredTitle(params.label) : params.label}
-        </span>
-        ${params.help ? html`<span class="settings-row__desc">${params.help}</span>` : nothing}
-      </label>
-      <div class="settings-row__control">${control}</div>
-    </div>
-  `;
-}
-
-type CronStringFormField = {
-  [Field in keyof CronFormState]: CronFormState[Field] extends string ? Field : never;
-}[keyof CronFormState];
-
-type CronBooleanFormField = {
-  [Field in keyof CronFormState]: CronFormState[Field] extends boolean ? Field : never;
-}[keyof CronFormState];
-
-type CronInputOptions = {
-  label: string;
-  help?: string;
-  placeholder?: string;
-  list?: string;
-  type?: string;
-  required?: boolean;
-  disabled?: boolean;
-  mono?: boolean;
-  errorKey?: CronFieldKey;
-  describeError?: boolean;
-};
-
-function renderCronInput(props: CronProps, field: CronStringFormField, options: CronInputOptions) {
-  const error = options.errorKey ? props.fieldErrors[options.errorKey] : undefined;
-  const describedBy =
-    error && options.errorKey && options.describeError !== false
-      ? errorIdForField(options.errorKey)
-      : undefined;
-  return html`
-    <input
-      id=${inputIdForField(field)}
-      class=${options.mono ? "settings-input mono" : "settings-input"}
-      type=${ifDefined(options.type)}
-      aria-required=${ifDefined(options.required ? "true" : undefined)}
-      .value=${props.form[field]}
-      list=${ifDefined(options.list)}
-      ?disabled=${options.disabled ?? false}
-      aria-invalid=${ifDefined(options.errorKey ? (error ? "true" : "false") : undefined)}
-      aria-describedby=${ifDefined(describedBy)}
-      placeholder=${ifDefined(options.placeholder)}
-      @input=${(event: Event) =>
-        props.onFormChange({ [field]: (event.currentTarget as HTMLInputElement).value })}
-    />
-  `;
-}
-
-function renderCronInputField(
-  props: CronProps,
-  field: CronStringFormField,
-  options: CronInputOptions,
-) {
-  const errorKey = options.errorKey;
-  return renderFieldRow({
-    label: options.label,
-    controlId: inputIdForField(field),
-    required: options.required,
-    help: options.help,
-    error: errorKey ? props.fieldErrors[errorKey] : undefined,
-    errorId: errorKey ? errorIdForField(errorKey) : undefined,
-    control: renderCronInput(props, field, options),
-  });
-}
-
-type CronSelectOptions = {
-  label: string;
-  options: readonly CronSelectOption[];
-  help?: string;
-  value?: string;
-  disabled?: boolean;
-  standalone?: boolean;
-  channel?: boolean;
-};
-
-function renderCronSelect(
-  props: CronProps,
-  field: CronStringFormField,
-  options: CronSelectOptions,
-) {
-  const selected = options.value ?? props.form[field];
-  const picker = options.channel ? renderChannelPicker : renderPicker;
-  return picker({
-    id: options.standalone ? undefined : inputIdForField(field),
-    label: options.label,
-    value: options.channel ? selected || "last" : selected,
-    options: options.options,
-    disabled: options.disabled,
-    onChange: (value) => props.onFormChange({ [field]: value }),
-  });
-}
-
-function renderCronSelectField(
-  props: CronProps,
-  field: CronStringFormField,
-  options: CronSelectOptions,
-) {
-  return renderFieldRow({
-    label: options.label,
-    controlId: inputIdForField(field),
-    help: options.help,
-    control: renderCronSelect(props, field, options),
-  });
-}
-
-function renderToggleRow(
-  props: CronProps,
-  field: CronBooleanFormField,
-  params: { label: string; help?: string },
-) {
-  return renderSettingsToggleRow({
-    title: params.label,
-    description: params.help,
-    checked: props.form[field],
-    onChange: (checked) => props.onFormChange({ [field]: checked }),
-  });
 }
 
 // ── Main render ──
